--- mobile/src/Poster/index.js
@@ -1,4 +1,0 @@
-/* @flow */
-import Poster from './presenter';
-
-export default Poster;
