+++ mobile/src/PosterScreen/index.js
@@ -1,0 +1,4 @@
+/* @flow */
+import PosterScreen from './presenter';
+
+export default PosterScreen;
