+++ mobile/src/PosterScreen/styles.js
